+++ changelog.md
@@ -17,9 +17,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Bug` for any issue/error occurred
 - `Installed` for installing new dependencies
 - `Setting` for adding/changing the general settings
-                        
+
+### [0.0.7] - 2021-12-13
+written by Hoin
+
+Installed
+ * msw                
+ * copy-webpack-plugin
+ * mui
+
+Setting
+ * mock service worker 환경 구축(https://imch.dev/posts/experiment-of-mock-service-worker)
+ * Webpack Config CopyPlugin 활용 mockServiceWorker.js build 시 카피
+
+Added
+ * addressApi => redux-toolkit 을 활용한 api mock 관 연동 및 활용
+ * AddressSearchComponent => 검색 된 장소를 서버와 연동 후 응답에 따른 액션 추가
+
+
+ 
 ### [0.0.5] - 2021-12-11
- written by hoin
+ written by Hoin
  Added
  * add Header.jsx
  * Installed/Setting - Webpack Env Config